refactor : reviewCnt 타입 수정

+++ src/main/java/com/sparta/first/project/eighteen/domain/stores/dtos/StoreResponseDto.java
@@ -39,7 +39,7 @@ public class StoreResponseDto {
 	private double storeRating;
 
 	// 식당 리뷰 개수
-	private int storeReviewCnt;
+	private long storeReviewCnt;
 
 	// 식당 이미지
 	private String storeImgUrl;
